Change way Java emedded in the Docker container When I trying to troubleshoot images, I have to build them. In the instructions there note that you should place `jre_8_32bit_noinstall.zip` in the `docker` folder. I initially go to the Oracle website in order to find latest Java executable, but did not find anything like that for Linux. There only jre-8u161-linux-i586.tar.gz for 32-bit version of JRE which I found. So I have to unpack TAR GZ archive and repackage it as ZIP in order to build. Which is not very obvious, and on Windows it will give not 100% same results as if image would be created in Linux (not interested for main usage, but possible solve some local issues).
From documentation standpoint would be probably better explicitly state that code from Oracle should be repackaged as ZIP archive, or replace code to use TAR GZ. and definitely provide information where to get archive required for packaging.

This is potentially could interfere with #35 , but it is better show
the code with explanation of idea behind change, then try to replicate
code with words.